refactor: mindre refaktorering
