+++ packages/popover-react/src/Popover.tsx
@@ -16,20 +16,20 @@ import {
     UseFloatingOptions,
 } from "@floating-ui/react";
 import { Props as FloatingFocusManagerOptions } from "@floating-ui/react/src/components/FloatingFocusManager";
-import { Props as UseClickOptions } from "@floating-ui/react/src/hooks/useClick";
-import { Props as UseDismissOptions } from "@floating-ui/react/src/hooks/useDismiss";
-import { Props as UseFocusOptions } from "@floating-ui/react/src/hooks/useFocus";
-import { Props as UseHoverOptions } from "@floating-ui/react/src/hooks/useHover";
-import { Props as UseRoleOptions } from "@floating-ui/react/src/hooks/useRole";
+import { Props as ClickOptions } from "@floating-ui/react/src/hooks/useClick";
+import { Props as DismissOptions } from "@floating-ui/react/src/hooks/useDismiss";
+import { Props as FocusOptions } from "@floating-ui/react/src/hooks/useFocus";
+import { Props as HoverOptions } from "@floating-ui/react/src/hooks/useHover";
+import { Props as RoleOptions } from "@floating-ui/react/src/hooks/useRole";
 import classNames from "classnames";
 import { merge } from "lodash";
 import * as React from "react";
-import { getThemeAndDensity } from "../../../portal/src/utils/getThemeAndDensity";
 import { Button } from "../../button-react";
+import { getThemeAndDensity } from "./utils";
 
 interface PopoverOptions {
     /**
-     * Initial open state of the popover.
+     * Initial open state of the popover when uncontrolled.
      * @default false
      * @see https://floating-ui.com/docs/useFloating#open
      */
@@ -46,37 +46,37 @@ interface PopoverOptions {
      * @see https://floating-ui.com/docs/useFloating#strategy
      */
     /** Options for hover
-     * @see {@link UseHoverOptions}
+     * @see {@link HoverOptions}
      * @see https://floating-ui.com/docs/useHover
      * @default { enabled: false }
      */
-    hoverOptions?: UseHoverOptions;
+    hoverOptions?: HoverOptions;
     /** Options for focus
-     * @see {@link UseFocusOptions}
+     * @see {@link FocusOptions}
      * @see https://floating-ui.com/docs/useFocus
      * @default { enabled: false }
      */
-    focusOptions?: UseFocusOptions;
+    focusOptions?: FocusOptions;
     /** Options for click
-     * @see {@link UseClickOptions}
+     * @see {@link ClickOptions}
      * @see https://floating-ui.com/docs/useClick
      * @default { enabled: false }
      */
-    clickOptions?: UseClickOptions;
+    clickOptions?: ClickOptions;
     /**
      * Options for role
-     * @see {@link UseRoleOptions}
+     * @see {@link RoleOptions}
      * @see https://floating-ui.com/docs/useRole
      * @default { enabled: true, role: "dialog" }
      */
-    roleOptions?: UseRoleOptions;
+    roleOptions?: RoleOptions;
     /**
      * Options for dismiss
-     * @see {@link UseDismissOptions}
+     * @see {@link DismissOptions}
      * @see https://floating-ui.com/docs
      * @default { enabled: false }
      */
-    dismissOptions?: UseDismissOptions;
+    dismissOptions?: DismissOptions;
     /**
      * Options for å posisjonere et flytende element
      * @see https://floating-ui.com/docs/useFloating
@@ -98,7 +98,7 @@ const usePopover = ({
     roleOptions,
     dismissOptions,
     floatingOptions: _floatingOptions,
-    floatingFocusManagerOptions: _floatingFocusManagerOptions,
+    floatingFocusManagerOptions,
 }: PopoverOptions) => {
     const [uncontrolledOpen, setUncontrolledOpen] = React.useState(initialOpen);
 
@@ -108,8 +108,8 @@ const usePopover = ({
     const floatingOptions = merge(
         {
             open,
-            placement: _floatingOptions?.placement ?? "bottom-start",
-            strategy: _floatingOptions?.strategy ?? "absolute",
+            placement: "bottom-start",
+            strategy: "absolute",
             middleware: [
                 offset(_offset),
                 flip({ padding: 5, fallbackPlacements: ["bottom", "top"] }),
@@ -121,38 +121,25 @@ const usePopover = ({
         _floatingOptions,
     );
 
-    const floatingFocusManagerOptions = merge(
-        {
-            order: ["reference", "floating", "content"],
-            initialFocus: 0,
-            guards: true,
-            returnFocus: true,
-            modal: true,
-            visuallyHiddenDismiss: true,
-            closeOnFocusOut: true,
-        },
-        _floatingFocusManagerOptions,
-    );
-
     const data = useFloating({ ...floatingOptions });
 
     const context = data.context;
 
     const role = useRole(context, {
+        enabled: true,
+        role: "dialog",
         ...roleOptions,
-        enabled: roleOptions?.enabled ?? true,
-        role: roleOptions?.role ?? "dialog",
     });
 
     const click = useClick(context, {
+        enabled: false,
         ...clickOptions,
-        enabled: (floatingOptions?.open || clickOptions?.enabled) ?? false,
     });
-    const hover = useHover(context, { ...hoverOptions, enabled: hoverOptions?.enabled ?? false });
-    const focus = useFocus(context, { ...focusOptions, enabled: focusOptions?.enabled ?? false });
+    const hover = useHover(context, { enabled: false, ...hoverOptions });
+    const focus = useFocus(context, { enabled: false, ...focusOptions });
     const dismiss = useDismiss(context, {
+        enabled: true,
         ...dismissOptions,
-        enabled: dismissOptions?.enabled ?? false,
     });
 
     const interactions = useInteractions([click, dismiss, focus, hover, , role]);
@@ -161,9 +148,9 @@ const usePopover = ({
         () => ({
             open,
             onOpenChange,
+            floatingFocusManagerOptions,
             ...interactions,
             ...data,
-            floatingFocusManagerOptions,
         }),
         [open, onOpenChange, interactions, data, floatingFocusManagerOptions],
     );
